+++ rest_framework_simple_api_key/models.py
@@ -1,36 +1,48 @@
+import typing
 from datetime import timedelta, datetime
 
+from django.conf import settings
 from django.db import models
 
 from rest_framework_simple_api_key.crypto import ApiKeyCrypto
-from rest_framework_simple_api_key.settings import package_settings
 
 
 def _expiry_date():
-    return datetime.now() + timedelta(days=360)
+    return datetime.now() + timedelta(settings.SIMPLE_API_KEY["API_KEY_LIFETIME"])
 
 
 class APIKeyManager(models.Manager):
-    key_handler = ApiKeyCrypto()
+    key_crypto = ApiKeyCrypto()
 
-    def get_usable_key(self):
-        pass
+    def get_key(self, pk: int | str):
+        return self.get(revoked=False, pk=pk)
 
-    def assign_key(self):
-        pass
+    def assign_key(self, obj) -> str:
+        payload = {"_pk": obj.pk, "_exp": obj.expiry_date.timestamp()}
+        key = self.key_crypto.generate(payload)
 
-    def create_key(self):
-        pass
+        return key
 
-    def revoke_api_keys(self):
-        pass
+    def create_key(self, **kwargs: typing.Any) -> typing.Tuple[typing.Any, str]:
+        # Prevent from manually setting the primary key.
+        obj = self.model(**kwargs)
+        obj.save()
+        key = self.assign_key(obj)
+
+        return obj, key
+
+    def revoke_api_key(self, pk: int | str):
+        api_key = self.get_key(pk)
+
+        api_key.revoked = True
+        api_key.save()
 
 
 class APIKey(models.Model):
     name = models.CharField(max_length=255, null=True, blank=True)
 
     entity = models.ForeignKey(
-        package_settings.SIMPLE_API_KEY["AUTHENTICATION_MODEL"],
+        settings.SIMPLE_API_KEY["AUTHENTICATION_MODEL"],
         on_delete=models.CASCADE,
         related_name="api_keys",
     )
@@ -38,13 +50,13 @@ class APIKey(models.Model):
     expiry_date = models.DateTimeField(
         default=_expiry_date,
         verbose_name="Expires",
-        help_text="Once API key expires, clients cannot use it anymore.",
+        help_text="Once API key expires, entities cannot use it anymore.",
     )
     revoked = models.BooleanField(
         blank=True,
         default=False,
         help_text=(
-            "If the API key is revoked, clients cannot use it anymore. "
+            "If the API key is revoked, entities cannot use it anymore. "
             "(This cannot be undone.)"
         ),
     )