+++ CHANGELOG.md
@@ -3,3 +3,5 @@ Changelog
 
 [Unreleased]
 ------------
+
+-  Add apikey model (#9)